+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 ### Security
+## [0.3.0] - 2020-09-01
 ## [0.3.0-7] - 2020-08-31
 
 ### Added
@@ -152,3 +153,4 @@ Initial prerelease.
 [0.3.0-5]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-4...v0.3.0-5
 [0.3.0-6]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-5...v0.3.0-6
 [0.3.0-7]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-6...v0.3.0-7
+[0.3.0]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-7...v0.3.0